test: move OpenClaw tests out of src
